[owfft] Move build_spec_table() to infrared/data.py
